Move snapshot-test helpers from snapshots/conftest.py into common/utils.py
``SNAPSHOT_SETTINGS``, ``EXPLAIN_SETTINGS`` and ``snapshot_given`` now
live alongside ``run_test_for_falsifying_example`` in
``tests/common/utils.py``, which is where the rest of the shared test
utilities already live. ``tests/snapshots/conftest.py`` no longer has
any content and is removed; the other ``tests/snapshots/test_*.py``
modules are updated to import from ``tests.common.utils``.
