+++ hypothesis/tests/common/utils.py
@@ -339,3 +339,56 @@ def run_test_for_falsifying_example(test_fn):
     with pytest.raises(AssertionError) as err:
         test_fn()
     return "\n".join(err.value.__notes__).strip()
+
+
+# --- Snapshot-test helpers -------------------------------------------
+#
+# These support the syrupy-based snapshot tests under tests/snapshots/.
+
+
+SNAPSHOT_SETTINGS = settings(
+    phases=[Phase.generate, Phase.shrink],
+    print_blob=False,
+    derandomize=True,
+    database=None,
+)
+
+EXPLAIN_SETTINGS = settings(
+    phases=[Phase.generate, Phase.shrink, Phase.explain],
+    print_blob=False,
+    derandomize=True,
+    database=None,
+)
+
+
+def snapshot_given(*strategies, **kwarg_strategies):
+    """Decorator that turns the wrapped body into a pytest test: runs it as
+    a Hypothesis property test and asserts that the captured
+    falsifying-example output equals the ``snapshot`` fixture value.
+
+    The body is expected to ``raise`` (e.g. ``AssertionError``) so the test
+    has a falsifying example to report. ``EXPLAIN_SETTINGS`` is used so
+    ``# or any other generated value``-style annotations participate in
+    the snapshot.
+    """
+    import functools
+    import inspect
+
+    from hypothesis import given
+
+    def decorator(body):
+        @functools.wraps(body)
+        def prop_body(*args, **kwargs):
+            body(*args, **kwargs)
+
+        prop_body.__signature__ = inspect.signature(body)
+        prop_test = given(*strategies, **kwarg_strategies)(EXPLAIN_SETTINGS(prop_body))
+
+        def test_function(snapshot):
+            assert run_test_for_falsifying_example(prop_test) == snapshot
+
+        test_function.__name__ = body.__name__
+        test_function.__qualname__ = body.__qualname__
+        return test_function
+
+    return decorator
