+++ hypothesis/tests/snapshots/test_combinators.py
@@ -12,8 +12,7 @@ from enum import Flag, auto
 
 from hypothesis import given, strategies as st
 
-from tests.common.utils import run_test_for_falsifying_example
-from tests.snapshots.conftest import SNAPSHOT_SETTINGS
+from tests.common.utils import SNAPSHOT_SETTINGS, run_test_for_falsifying_example
 
 
 class Direction(Flag):
