+++ hypothesis/tests/snapshots/test_explain.py
@@ -10,8 +10,11 @@
 
 from hypothesis import given, strategies as st
 
-from tests.common.utils import run_test_for_falsifying_example
-from tests.snapshots.conftest import EXPLAIN_SETTINGS
+from tests.common.utils import (
+    EXPLAIN_SETTINGS,
+    run_test_for_falsifying_example,
+    snapshot_given,
+)
 
 
 def test_explain_comments_basic_fail_if_either(snapshot):
@@ -169,3 +172,18 @@ def test_explain_unstable_one_of_labels(snapshot):
         assert not isinstance(w.a, list)
 
     assert run_test_for_falsifying_example(inner) == snapshot
+
+
+@snapshot_given(st.integers(), st.integers())
+def test_integers_equal(n1, n2):
+    assert n1 == n2
+
+
+@snapshot_given(st.integers(), st.integers())
+def test_integers_less(n1, n2):
+    assert n1 < n2
+
+
+@snapshot_given(st.integers(), st.integers())
+def test_integers_greater(n1, n2):
+    assert n1 < n2
